drop ApiConfig.java
