+++ api/src/main/java/org/accula/api/config/WebConfig.java
@@ -1,13 +1,21 @@
 package org.accula.api.config;
 
 import lombok.RequiredArgsConstructor;
+import org.accula.api.code.RepositoryManager;
+import org.accula.api.code.RepositoryProvider;
 import org.accula.api.db.repo.CurrentUserRepo;
+import org.accula.api.detector.CloneDetector;
+import org.accula.api.detector.PrimitiveCloneDetector;
 import org.accula.api.github.api.GithubClient;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringBootConfiguration;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.web.reactive.config.WebFluxConfigurer;
 import org.springframework.web.reactive.function.client.WebClient;
 import reactor.core.publisher.Mono;
+
+import java.io.File;
 
 /**
  * @author Anton Lamtev
@@ -15,6 +23,7 @@ import reactor.core.publisher.Mono;
  */
 @SpringBootConfiguration
 @RequiredArgsConstructor
+@EnableConfigurationProperties(WebhookProperties.class)
 public class WebConfig implements WebFluxConfigurer {
     private final CurrentUserRepo currentUserRepository;
 
@@ -35,5 +44,16 @@ public class WebConfig implements WebFluxConfigurer {
         return () -> currentUserRepository
                 .get()
                 .flatMap(user -> Mono.just(user.getGithubUser().getLogin()));
+    }
+
+    @Bean
+    public RepositoryProvider repositoryManager(@Value("${accula.reposPath}") final String reposPath) {
+        final File root = new File(reposPath);
+        return new RepositoryManager(root);
+    }
+
+    @Bean
+    public CloneDetector cloneDetector() {
+        return new PrimitiveCloneDetector(1, 2);
     }
 }
